+++ tests/test_voxelization.py
@@ -1,37 +1,31 @@
 import unittest
 import numpy as np
 
-from sigmaepsilon.mesh import PolyData, PointData
-from sigmaepsilon.mesh.space import StandardFrame
-from sigmaepsilon.mesh.cells import H8
-
-from sigmaepsilon.mesh.voxelize import voxelize_TET4_H8
+from sigmaepsilon.mesh.voxelize import voxelize_TET4_H8, voxelize_T3_H8
 
 
-class Test_voxelize_TET4_H8(unittest.TestCase):
-    def test_1(self):
+class TestVoxelize(unittest.TestCase):
+    def test_voxelize_TET4_H8_1__SMOKE(self):
         coords_TET4 = np.array([[0, 0, 0], [1, 0, 0], [0, 1, 0], [0, 0, 1]]).astype(
             float
         )
         topo_TET4 = np.array([[0, 1, 2, 3]])
         shape = (10, 10, 10)
-        hexa_mesh = voxelize_TET4_H8(coords_TET4, topo_TET4, shape=shape, k_max=1)
+        voxelize_TET4_H8(coords_TET4, topo_TET4, shape=shape, k_max=1)
 
-        frame = StandardFrame(dim=3)
-        mesh = PolyData(
-            pd=PointData(coords=hexa_mesh[0], frame=frame),
-            cd=H8(topo=hexa_mesh[1], frames=frame),
-        )
-
-        self.assertTrue(np.isclose(mesh.volume(), 0.2829999999999999))
-
-    def test_2(self):
+    def test_voxelize_TET4_H8_2__SMOKE(self):
         coords_TET4 = np.array([[0, 0, 0], [1, 0, 0], [0, 1, 0], [0, 0, 1]]).astype(
             float
         )
         topo_TET4 = np.array([[0, 1, 2, 3]])
         resolution = 0.1
         voxelize_TET4_H8(coords_TET4, topo_TET4, resolution=resolution, k_max=1)
+
+    def test_voxelize_T3_H8_1__SMOKE(self):
+        coords_T3 = np.array([[0, 0, 0], [1, 0, 0], [0, 1, 0]]).astype(float)
+        topo_T3 = np.array([[0, 1, 2]])
+        shape = (10, 10, 10)
+        voxelize_T3_H8(coords_T3, topo_T3, shape=shape, k_max=10)
 
 
 if __name__ == "__main__":
